WIP: Draft email with clearification on how Moore grant can be spend
The finance comitee is looking at the Moore grant application. The application
lays out a budget, but it seems that the rules for spending are not as strict
as some of us are used to from e.g. NASA grants. However, its always better
to be upfront and ask early, rather than to find out later that the deadline for
spending money has already passed and money has to be returned.

Thus, we plan to contact the Moore foundation to ask for the rules.

**Deadline: Feb 15 2020**

This PR will be mereged and the email send after the deadline, so any comments
have to be done before that date.

closes #47