funding_requests.md Revisions
I've made mostly grammatical and proposal changes though I did add some information that should help to facilitate the workflow between the Finance Committee and NumFOCUS. Note that "Interim Finance Committee" was not used consistently in the original file so I changed all mentions to "Finance Committee". Though this might not be entirely accurate at the moment it could save some work down the road to leave out the "interim". I think it's clear to the community who you are referring to. A couple of things that are optional but may improve the policy:

1. Consider having questions about this policy submitted as an issue rather than an email to finance@astropy.org and directing those with questions to first check previous issues. This may save the committee some time by avoiding duplicate questions.

2. Consider including a short list of items that are typically approved and not approved for payment. This again cuts down on questions but also gives the committee an explicit and transparent way to justify rejecting certain requests and accepting others.