Draft a policy on email lists and slack channels
The goal of this policy is to write down how mailing lists (both public and restricted) are set up in astropy. Who gets to be a member? Is the content available openly? Are the lists moderated?

This is a WIP for several reasons:

- This not an "as-is" (e.g. I don't know if any of our mailing lists are currently moderated), but an "as I want it to be" that needds discussion and community consensus.
- I may not be aware of all communication channels used. Missing channels need to be added (with membership restriction as appropriate) - please add comments or PRs to my branch.

I think this is a policy that's way below the level of e.g. APE0, thus I think it's appropriate to suggest this as a PR in the project repository. Like an APE, it should probably annouced on astropy-dev, when the list of lists is more complete.