Formalize period for improvements of affiliated packages
As written, the rules for affilaited pacakges give only one month of time
to resposnd to the review of an affilaited package; after that time, the
editor has no choice but to close the PR as rejected.
One month is sure sufficient to reply to a review, but not enough to implement
any changes. We know that even in astropy itself, if often takes well over a
months to open a PR, review, and merge it.
This leaves editors with two bad options: (1) Reject a package that is almost,
but not quite ready to be accepted. While a the package author can in theory
apply again at a later point, it would be a lot more friendly to wait a little
longer to give time to implement changesm or (2) accept a packages that is
almost but not quite ready for the sake of including and promoting all
community work and "to put it out there".

Additional thoughts:

- Often, improvments requested are for better integration
  with astropy (quantities, Spectrum1D) - changes that are possible but deep
  enough that no package wants to rush them through or documentation.
- In principle, we can re-review packages, but in practice we have not done
  that ever due to limited resources, so the initial review is really the best
  point in time to motivate an affilated package to improve certain areas.