Cycle 4 Funding: Community Engagement Follow-on work by OrgMycology
We've attached a proposal to create a follow-on for the work we've been doing with Astropy. We see some key areas we think we could be helpful going forward, and welcome conversation over the coming weeks to co-create some intitiatives under this outline proposal. We'd very much like to hear from the CoCo about priorities they might have from our community survey that they think we could potentially contribute to.